reshape tests + remove unused functions
